Added another image in the demo app
newton-cradle.gif is an image with a high frame rate (35 frames in less
than a second), which makes it interesting for performance tests.
Initially, frames weren't rendered fast enough, and some frames were
skipped, causing rendering issues. I switched to using TimingManager
instead of an animation, so frames are not skipped anymore, but it's still
not fast enough.